Rename and move TransactionParts
