+++ notary-plugins/notary-plugin-non-validating/notary-plugin-non-validating-server/src/main/kotlin/com/r3/corda/notary/plugin/nonvalidating/NonValidatingNotaryTransactionDetails.kt
@@ -1,4 +1,4 @@
-package com.r3.corda.notary.plugin.common
+package com.r3.corda.notary.plugin.nonvalidating
 
 import net.corda.v5.crypto.SecureHash
 import net.corda.v5.ledger.utxo.StateRef
@@ -8,7 +8,7 @@ import net.corda.v5.ledger.utxo.TimeWindow
  * The minimum amount of information needed to notarise a transaction. Note that this does not include
  * any sensitive transaction details.
  */
-data class TransactionParts(
+data class NonValidatingNotaryTransactionDetails(
     val id: SecureHash,
     val numOutputs: Int,
     val timeWindow: TimeWindow,
